Add lib/process-files; refactor dotfiles, user-bin
The `process_files` functional abstraction makes it easy to process
files in a uniform way without `for` or `while` loops or external
programs like `find`. I'll probably move it into go-script-bash
eventually, along with a similar `process_dirs` and `process_all`.

I'm also about to use it to implement the `diff` command, which will
allow me to see how I've modified dotfiles either on the host or in the
repo, and consider porting the changes one way or the other.