Enable successful install on macOS 10.12.6
Of particular interest are the updates to `./go install languages go`
whereby the Go source repo is primed with the `GOROOT_BOOTSTRAP` version
in `_prepare_go_repo_for_bootstrap`. This is necessary due to the macOS
Sierra issues described in https://github.com/golang/go/issues/16352.
The fix is only in the `release-branch.go1.4` branch not a tag, which
causes `gvm` not to find it without this priming step.